Update unique_granges.R
I had a need to remove duplicate columns from GRanges that were joined (using plyranges::join_overlap_intersect). The issue was that a suffix ".1" is added to shared metadata, so I have included the lines to assess for duplicated colnames with that suffix and remove them by dplyr::select where extant.